feat:location ubah database dan tambah alert

+++ resources/views/pages/data-master/location/_function/function.blade.php
@@ -25,18 +25,39 @@
         });
     }
 
+
     function deleteFunc(id) {
         var id = id;
-        $.ajax({
-            type: "DELETE",
-            url: "{{ route('location.destroy') }}",
-            data: {
-                id: id
-            },
-            dataType: 'json',
-            success: function(res) {
-                var oTable = $('#myTable').dataTable();
-                oTable.fnDraw(false);
+        Swal.fire({
+            title: 'Apakah Anda yakin?',
+            text: "Data yang dihapus tidak dapat dikembalikan!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#d33',
+            cancelButtonColor: '#3085d6',
+            confirmButtonText: 'Ya, Hapus!',
+            cancelButtonText: 'Batal'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                $.ajax({
+                    type: "DELETE",
+                    url: "{{ route('location.destroy') }}",
+                    data: {
+                        id: id
+                    },
+                    dataType: 'json',
+                    success: function(res) {
+                        var oTable = $('#myTable').dataTable();
+                        oTable.fnDraw(false);
+                        Swal.fire({
+                            icon: 'success',
+                            title: 'Success',
+                            text: res.message,
+                            timer: 2000,
+                            showConfirmButton: false
+                        })
+                    }
+                });
             }
         });
     }
@@ -58,9 +79,36 @@
                 oTable.fnDraw(false);
                 $("#btn-save").html('Submit');
                 $("#btn-save").attr("disabled", false);
+                if (response.success == true) {
+                    Swal.fire({
+                        icon: 'success',
+                        title: 'Success',
+                        text: response.message,
+                        timer: 1000,
+                        showConfirmButton: false
+                    });
+                } else if (response.success == false) {
+                    Swal.fire({
+                        icon: 'error',
+                        title: 'Failed',
+                        text: response.message,
+                        timer: 2000,
+                        showConfirmButton: false
+                    });
+                }
             },
-            error: function(data) {
-                console.log(data);
+            error: function(xhr, status, error) {
+                var errorMessage = xhr.status + ': ' + xhr.statusText;
+                console.error('Error - ' + errorMessage);
+
+                // Sweet Alert for Error
+                // Swal.fire({
+                //     icon: 'error',
+                //     title: 'Error',
+                //     text: 'Failed to save data. Please try again later.',
+                //     timer: 2000,
+                //     showConfirmButton: false
+                // });
             }
         });
     });
